fix(models): allow float for price_from / subscribers
Haiku parses site prices like "6.4 млн" as 6.4 (float). Pydantic v2
rejects float→int with fractional part, so SiteFacts validation failed
mid-pipeline. Switched ProductRef.price_from, ProjectRef.price_from,
SocialChannel.subscribers from int to float — int values still flow
through fine.

Bug spotted during first cold-intro run on a real client site.

+++ src/models.py
@@ -92,7 +92,7 @@ class IntroPair(BaseModel):
 class ProductRef(BaseModel):
     name: str
     url: str | None = None
-    price_from: int | None = None
+    price_from: float | None = None
     description: str | None = None
 
 
@@ -100,7 +100,7 @@ class ProjectRef(BaseModel):
     name: str
     url: str | None = None
     location: str | None = None
-    price_from: int | None = None
+    price_from: float | None = None
     status: str | None = None
 
 
@@ -126,7 +126,7 @@ class SocialChannel(BaseModel):
     ]
     handle: str | None = None
     url: str | None = None
-    subscribers: int | None = None
+    subscribers: float | None = None
 
 
 class SiteFacts(BaseModel):
